Remove #children fron BaseEvent (#90)
`#children` has been deprecated as of Rails 7.2.

https://github.com/rails/rails/pull/43390

Unlike Rails, I decided to remove `#children` from `BaseEvent`
since it'd be a techdebt to keep this method in the future,
and I guess this method would not be used very well.

By the way, I found this deprecation on Rails when I'm about to
add the support for `deprecation.rails` on #89. I hope this won't cause
problems for users who use this gem.