Add `:db_runtime` to `perform.active_job`
`:db_runtime` was added in the payload of `perform.active_job`,
so make this gem support it.

I also updated the version condition to use `>= Gem::Version.new('7.1.0.alpha')`
instead of `> Gem::Version.new('7.0')`. I think it would be better to
specify the version that supports the feature, not the version doesn't
support it.